Used the Easystar.js library to find/trace shortest path from test enemy to player character upon pressing 'ctrl' (A_BUTTON_DOWN).
As a proof of concept, a new Enemy sprite is created and placed at each cell in the path found by the A Star pathfinding algorithm.

At this point, the player can move his/her character with the WASD keys or for debugging purposes, click on any cell to move there.